// src/lib/api/base/api.ts
export interface AppSchemasStrategyProfitAndLoss {
    /**
     * 
     * @type {any}
     * @memberof AppSchemasStrategyProfitAndLoss
     */
    'strategy_direction'?: any;
    /**
     * 
     * @type {any}
     * @memberof AppSchemasStrategyProfitAndLoss
     */
    'max_profit'?: any;
    /**
     * 
     * @type {any}
     * @memberof AppSchemasStrategyProfitAndLoss
     */
    'max_loss'?: any;
    /**
     * 
     * @type {any}
     * @memberof AppSchemasStrategyProfitAndLoss
     */
    'breakeven_points': any;
    /**
     * 
     * @type {any}
     * @memberof AppSchemasStrategyProfitAndLoss
     */
    'pop'?: any;
}
// ...
export interface OptionChainParsedResult {
    /**
     * 
     * @type {any}
     * @memberof OptionChainParsedResult
     */
    'index_name': any;
    /**
     * 
     * @type {any}
     * @memberof OptionChainParsedResult
     */
    'time'?: any;
    /**
     * 
     * @type {any}
     * @memberof OptionChainParsedResult
     */
    'expiry_date': any;
    /**
     * 
     * @type {any}
     * @memberof OptionChainParsedResult
     */
    'spot_price_round_off': any;
    /**
     * 
     * @type {any}
     * @memberof OptionChainParsedResult
     */
    'spot_price': any;
    /**
     * 
     * @type {any}
     * @memberof OptionChainParsedResult
     */
    'option_chain_data': any;
}
// ...
export interface PlotResult {
    /**
     * 
     * @type {any}
     * @memberof PlotResult
     */
    'result'?: any;
    /**
     * 
     * @type {any}
     * @memberof PlotResult
     */
    'spot_prices'?: any;
    /**
     * 
     * @type {any}
     * @memberof PlotResult
     */
    'strike_prices'?: any;
}
// ...
export interface StrategyPayOffData {
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'strategy_data'?: any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'spot_price': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'spot_price_entered'?: any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'spot_price_diff'?: any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'index_name': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'expiry_date': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'time': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'strategy_name': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'total_value': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'strategy_journey': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'strategy_exit_journey': any;
    /**
     * 
     * @type {AppSchemasStrategyProfitAndLoss}
     * @memberof StrategyPayOffData
     */
    'profit_and_loss': AppSchemasStrategyProfitAndLoss;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'script_lot_size': any;
    /**
     * 
     * @type {PlotResult}
     * @memberof StrategyPayOffData
     */
    'plot_results': PlotResult;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'pop'?: any;
}